+++ Fun.Build/StageBuilder.fs
@@ -11,15 +11,42 @@ type StageBuilder(name: string) =
 
 
     member inline _.Yield(_: unit) = BuildStage id
-
-    member inline _.Yield([<InlineIfLambda>] condition: BuildStageIsActive) = condition
-    member inline _.Yield([<InlineIfLambda>] builder: BuildStep) = builder
     member inline _.Yield(stage: StageContext) = stage
-
 
     member inline _.Delay([<InlineIfLambda>] fn: unit -> BuildStage) = BuildStage(fun ctx -> fn().Invoke(ctx))
 
-    member inline _.Delay([<InlineIfLambda>] fn: unit -> BuildStageIsActive) = BuildStage(fun ctx -> { ctx with IsActive = fn().Invoke })
+    member inline _.Delay([<InlineIfLambda>] fn: unit -> StageContext) =
+        BuildStage(fun ctx -> { ctx with Steps = ctx.Steps @ [ Step.StepOfStage(fn ()) ] })
+
+    member inline _.Combine(stage: StageContext, [<InlineIfLambda>] build: BuildStage) =
+        BuildStage(fun ctx -> build.Invoke { ctx with Steps = ctx.Steps @ [ Step.StepOfStage stage ] })
+
+    member inline _.For([<InlineIfLambda>] build: BuildStage, [<InlineIfLambda>] fn: unit -> BuildStage) =
+        BuildStage(fun ctx -> fn().Invoke(build.Invoke(ctx)))
+
+    member inline _.For([<InlineIfLambda>] build: BuildStage, [<InlineIfLambda>] fn: unit -> StageContext) =
+        BuildStage(fun ctx ->
+            let ctx = build.Invoke ctx
+            { ctx with Steps = ctx.Steps @ [ Step.StepOfStage(fn ()) ] }
+        )
+
+
+    member inline _.Yield([<InlineIfLambda>] condition: BuildStageIsActive) = condition
+
+    member inline _.Delay([<InlineIfLambda>] fn: unit -> BuildStageIsActive) =
+        BuildStage(fun ctx -> { ctx with IsActive = fun ctx -> fn().Invoke(ctx) })
+
+    member inline _.Combine([<InlineIfLambda>] condition: BuildStageIsActive, [<InlineIfLambda>] build: BuildStage) =
+        BuildStage(fun ctx -> build.Invoke { ctx with IsActive = condition.Invoke })
+
+    member inline _.For([<InlineIfLambda>] build: BuildStage, [<InlineIfLambda>] fn: unit -> BuildStageIsActive) =
+        BuildStage(fun ctx ->
+            let ctx = build.Invoke ctx
+            { ctx with IsActive = fun ctx -> fn().Invoke(ctx) }
+        )
+
+
+    member inline _.Yield([<InlineIfLambda>] builder: BuildStep) = builder
 
     member inline _.Delay([<InlineIfLambda>] fn: unit -> BuildStep) =
         BuildStage(fun ctx ->
@@ -28,32 +55,12 @@ type StageBuilder(name: string) =
             }
         )
 
-    member inline _.Delay([<InlineIfLambda>] fn: unit -> StageContext) =
-        BuildStage(fun ctx -> { ctx with Steps = ctx.Steps @ [ Step.StepOfStage(fn ()) ] })
-
-
-    member inline _.Combine([<InlineIfLambda>] condition: BuildStageIsActive, [<InlineIfLambda>] build: BuildStage) =
-        BuildStage(fun ctx -> build.Invoke { ctx with IsActive = condition.Invoke })
-
     member inline _.Combine([<InlineIfLambda>] builder: BuildStep, [<InlineIfLambda>] build: BuildStage) =
         BuildStage(fun ctx ->
             build.Invoke
                 { ctx with
                     Steps = ctx.Steps @ [ Step.StepFn builder.Invoke ]
                 }
-        )
-
-    member inline _.Combine(stage: StageContext, [<InlineIfLambda>] build: BuildStage) =
-        BuildStage(fun ctx -> build.Invoke { ctx with Steps = ctx.Steps @ [ Step.StepOfStage stage ] })
-
-
-    member inline _.For([<InlineIfLambda>] build: BuildStage, [<InlineIfLambda>] fn: unit -> BuildStage) =
-        BuildStage(fun ctx -> fn().Invoke(build.Invoke(ctx)))
-
-    member inline _.For([<InlineIfLambda>] build: BuildStage, [<InlineIfLambda>] fn: unit -> BuildStageIsActive) =
-        BuildStage(fun ctx ->
-            let ctx = build.Invoke ctx
-            { ctx with IsActive = fn().Invoke }
         )
 
     member inline _.For([<InlineIfLambda>] build: BuildStage, [<InlineIfLambda>] fn: unit -> BuildStep) =
@@ -64,11 +71,6 @@ type StageBuilder(name: string) =
             }
         )
 
-    member inline _.For([<InlineIfLambda>] build: BuildStage, [<InlineIfLambda>] fn: unit -> StageContext) =
-        BuildStage(fun ctx ->
-            let ctx = build.Invoke ctx
-            { ctx with Steps = ctx.Steps @ [ Step.StepOfStage(fn ()) ] }
-        )
 
     /// Add or override environment variables
     [<CustomOperation("envVars")>]